pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "SEBI Hackathon - Investor Education Platform Backend"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "alembic>=1.12.1",
    "psycopg2-binary>=2.9.9",
    "redis>=5.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
    "langchain>=0.1.0",
    "langchain-openai>=0.0.5",
    "langchain-community>=0.0.10",
    "langchain-experimental>=0.0.49",
    "langgraph>=0.0.20",
    "pinecone-client>=3.0.0",
    "pgvector>=0.2.4",
    "faiss-cpu>=1.7.4",
    "sentence-transformers>=2.2.2",
    "instructorembedding>=1.0.1",
    "livekit>=0.11.0",
    "livekit-agents>=0.8.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
    "opentelemetry-instrumentation-fastapi>=0.42b0",
    "opentelemetry-exporter-otlp>=1.21.0",
    "prometheus-client>=0.19.0",
    "structlog>=23.2.0",
    "httpx>=0.25.2",
    "aioredis>=2.0.1",
    "pypdf2>=3.0.1",
    "beautifulsoup4>=4.12.2",
    "numpy>=1.26.2",
    "pandas>=2.1.3",
    "python-dotenv>=1.0.0",
    "langchain-text-splitters>=0.3.9",
    "aiohttp>=3.9.0",
    "pymupdf>=1.26.3",
    "langchain-chroma>=0.2.5",
    "pymongo[srv]>=4.14.1",
    "langchain-anthropic>=0.3.19",
    "uuid>=1.30",
    "langchain-google-genai>=2.1.9",
    "langgraph-cli[inmem]>=0.3.8",
    "motor>=3.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "isort>=5.12.0",
    "mypy>=1.7.1",
    "pre-commit>=3.6.0"
]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
